+++ test/specs/parsers/_suite.js
@@ -8,20 +8,18 @@
 var
   path = require('path'),
   fs   = require('fs'),
-  norm    = require('../helpers').normalizeJS,
-  reqname = require('../helpers').requireName
+  norm = require('../helpers').normalizeJS
 
 var
   fixtures = __dirname,
   expected = path.join(fixtures, 'js'),
   parsers  = compiler.parsers
 
-function have (mod, req) {
-  if (parsers._req(mod, req)) return true
+function have (name) {
+  if (parsers._req(name)) return true
 
-  if (mod !== 'unknown') {
-    if (!req) req = reqname(mod)
-    console.error('\tnot installed locally: ' + req + ' alias "' + mod + '"')
+  if (name !== 'unknown') {
+    console.error('\t' + name + ' parser not installed locally.')
   }
   return false
 }
@@ -59,6 +57,12 @@ describe('HTML parsers', function () {
     if (have('jade') && have('coffee')) {
       testParser('test.jade', { template: 'jade' })
       testParser('slide.jade', { template: 'jade' })
+    }
+  })
+  it('pug', function () {
+    if (have('pug') && have('coffee')) {
+      testParser('test.pug', { template: 'pug' })
+      testParser('slide.pug', { template: 'pug' })
     }
   })
 
@@ -240,7 +244,7 @@ describe('Style parsers', function () {
 
   // scss.tag
   it('custom parser using postcss + autoprefixer', function () {
-    if (have('postcss', 'postcss')) {
+    if (have('postcss')) {
       testParser('postcss')
     }
   })
@@ -273,6 +277,32 @@ describe('Style parsers', function () {
     expect(result).to.contain('P {top:0}')
   })
 
+  it('the style parser options can be passed directly to the compiler', function () {
+    var
+      source = [
+        '<style-option>',
+        '  <style>',
+        '    p {top:0}',
+        '  </style>',
+        '</style-option>'
+      ].join('\n')
+
+    parsers.css.myParser2 = function (tag, css, opts) {
+      expect(opts.flag).to.be(true)
+      // don't do anything
+      return css
+    }
+
+    compiler.compile(source, {
+      style: 'myParser2',
+      parserOptions: {
+        style: {
+          flag: true
+        }
+      }
+    })
+  })
+
 })
 
 describe('Other', function () {
@@ -296,6 +326,56 @@ describe('Other', function () {
     expect(compiler.compile).withArgs(str1, { type: 'unknown' }).to.throwError()
     expect(compiler.compile).withArgs(str2).to.throwError()
     expect(have('unknown')).to.be(false)
+  })
+
+  it('the js parser options can be passed directly to the compiler', function () {
+    var
+      tag = [
+        '<custom-options>',
+        'this.foo = () => null',
+        '<custom-options>'
+      ].join('\n')
+
+    parsers.js.foo = function (js, opts) {
+      expect(opts.flag).to.be(true)
+      // don't do anything
+      return js
+    }
+
+    compiler.compile(tag, {
+      js: 'foo',
+      parserOptions: {
+        js: {
+          flag: true
+        }
+      }
+    })
+
+  })
+
+  it('the template parser options can be passed directly to the compiler', function () {
+    var
+      tag = [
+        '<custom-options>',
+        'this.foo = () => null',
+        '<custom-options>'
+      ].join('\n')
+
+    parsers.html.foo = function (html, opts) {
+      expect(opts.flag).to.be(true)
+      // don't do anything
+      return ''
+    }
+
+    expect(compiler.compile(tag, {
+      template: 'foo',
+      parserOptions: {
+        template: {
+          flag: true
+        }
+      }
+    })).to.be('')
+
   })
 
   // brackets.tag